many changes, but now it can interpolate and restore reports (except wind direction column)

+++ src/main/java/com/kpi/tefresolver/model/ObservationData.java
@@ -23,7 +23,7 @@ public class ObservationData {
 
     @Column
 //    @Size(min = 1, max = 31)
-    private Short dayNumber;
+    private Integer dayNumber;
 
     @Column
 //    @Temporal(TemporalType.TIME)
@@ -31,7 +31,7 @@ public class ObservationData {
 
     @Column
 //    @Min(-273)
-    private Short temperature;
+    private Integer temperature;
 
     @Column
 //    @Enumerated(EnumType.STRING)
